+++ src/components/exercises/Sortable/Sortable.jsx
@@ -1,11 +1,8 @@
 // Sortable.jsx
-import {
-	AudioClip,
-	IconButton,
-	Info,
-	ProgressDots,
-} from "@/components";
 import { exerciseActionButtonVariants } from "@/components/exercises/shared/exerciseActionButtonVariants";
+import { ProgressDots } from "@/components/exercises/ProgressDots";
+import { Info } from "@/components/content";
+import { AudioClip, IconButton } from "@/components/media";
 import React from "react";
 import { SortableWordCard } from "@/components/exercises/SortableWordCard/SortableWordCard";
 import { captureFlipPositions, playFlipAnimation } from "@/utils/reorderAnimation";